requirements.txt
dvc[gdrive]==2.34.1
ipywidgets==8.0.2
jupyterlab==3.5.0
matplotlib==3.6.0
numpy==1.23.4
omegaconf==2.3.0
pandas==1.5.1
pillow==9.4.0
pytorch_lightning==1.8.0
scikit-learn==1.1.3
seaborn==0.12.1
torch==1.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
wandb==0.13.5
